Use percentage of area instead of radius to display battery remaining.
Currently the radius of the inner battery circle is calculated by multiplying
it by the battery fraction directly. This results in the circle looking
smaller than expected because the area is smaller.

For example, when the battery is at 50% the battery fraction is 0.5.
The area of the outer circle is Pi*r.
The area of the inner circle is Pi*r*(0.5)^2 which is Pi*r*0.25.
So the area of the inner circle is 25% of the outer circle instead of 50%.

This changes the calculation so that the inner circle area reflects the battery
percentage instead of the radius.

Tested on a Nexus 5X running Marshmellow.